# docs/swagger.yaml
definitions:
...
  helper.ErrorResponse:
    properties:
      err:
        type: string
      message:
        type: string
    type: object
  helper.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.CommentReqEq:
    properties:
      message:
        example: Ambis Sekali
        type: string
      photo_id:
        example: 1
        type: integer
    type: object
...
  model.PhotoReqEq:
    properties:
      caption:
        example: Salah satu SS challange-09
        type: string
      photo_url:
        example: https://res.cloudinary.com/drakr4vtu/image/upload/v1681832825/FGA%20GO%2010/challange-09/image_2023-04-18_224711118_xcyqhz.png
        type: string
      title:
        example: ch-09 10
        type: string
    type: object
...
  model.UpdateCommentReq:
    properties:
      message:
        example: Ambis Sekali yah bund
        type: string
      photo_id:
        example: 1
        type: integer
    type: object
  model.UpdatePhotoReq:
    properties:
      caption:
        example: Salah satu SrenShot challange-09
        type: string
      photo_url:
        example: https://res.cloudinary.com/drakr4vtu/image/upload/v1681832825/FGA%20GO%2010/challange-09/image_2023-04-18_224711118_xcyqhz.png
        type: string
      title:
        example: ch-09 10-0
        type: string
    type: object
...
info:
  contact: {}
paths:
  /comment:
    get:
      consumes:
      - application/json
      description: Get all comment
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all comment
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create a new Comment
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.CommentReqEq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new Comment
      tags:
      - Comment
  /comment/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by ID
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a comment by ID
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: Get comment by ID
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comment by ID
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update existing comment
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: comment ID
        in: path
        name: commentId
        required: true
        type: integer
      - description: Update Comment object
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update existing comment
      tags:
      - Comment
  /photo:
    get:
      consumes:
      - application/json
      description: Get all photo
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all photo
      tags:
      - Photo
  /photo/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete a photo by ID
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a photo by ID
      tags:
      - Photo
    get:
      consumes:
      - application/json
      description: Get photo by ID
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: photo ID
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get photo by ID
      tags:
      - Photo
    put:
      consumes:
      - application/json
      description: Update existing photo
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      - description: Update photo object
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePhotoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update existing photo
      tags:
      - Photo
  /photos:
    post:
      consumes:
      - application/json
      description: Create a new Photo
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sosmed data
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/model.PhotoReqEq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new Photo
      tags:
      - Photo
...
swagger: "2.0"